+++ CHANGELOG.md
@@ -11,9 +11,10 @@ Package  | Version  | Link
 ---|---|---
 SDK Core| v1.0.3 | [symbol-sdk](https://www.npmjs.com/package/symbol-sdk)
 Catbuffer | v1.0.1 | [catbuffer-typescript](https://www.npmjs.com/package/catbuffer-typescript)
-Client Library | v1.0.2  | [symbol-openapi-typescript-fetch-client](https://www.npmjs.com/package/symbol-openapi-typescript-fetch-client)
+Client Library | v1.0.3  | [symbol-openapi-typescript-fetch-client](https://www.npmjs.com/package/symbol-openapi-typescript-fetch-client)
 
 - fix: replaced `instanceof` statements. These statements are problematic when npm installs the dependency in multiples modules. 
+- feat: added mosaic revocation support.
 
 ## [1.0.2] - 25-Oct-2021
 
@@ -21,7 +22,7 @@ Client Library | v1.0.2  | [symbol-openapi-typescript-fetch-client](https://www.
 Package  | Version  | Link
 ---|---|---
 SDK Core| v1.0.2 | [symbol-sdk](https://www.npmjs.com/package/symbol-sdk)
-Catbuffer | v1.0.1 | [catbuffer-typescript](https://www.npmjs.com/package/catbuffer-typescript)
+Catbuffer | v1.0.0 | [catbuffer-typescript](https://www.npmjs.com/package/catbuffer-typescript)
 Client Library | v1.0.2 | [symbol-openapi-typescript-fetch-client](https://www.npmjs.com/package/symbol-openapi-typescript-fetch-client)
 
 - feat: Multisig multilevel subscription in web listener.